refactor : move all `store.ts` files into `index.ts`
