+++ src/hooks/mutations/useRemoveFilmRecipeMutation.ts
@@ -3,7 +3,7 @@ import { removeRecipe } from '../../supabase/filmRecipe';
 import { RestrictedRecipe } from '../../supabase/schema';
 import queryKey from '../../constants/queryKey';
 import { toastData } from '../../constants/toast';
-import useToastStore from '../../store/useToastStore';
+import { useToastStore } from '../../store';
 
 interface UseRemoveFilmRecipeMutation {
 	id: string;
