+++ src/pages/ExpenseTrackerByMonth.tsx
@@ -1,7 +1,7 @@
 import { Suspense, useState } from 'react';
 import styled from '@emotion/styled';
 import { DatePicker, PaymentItemLoader, PaymentList, SegmentedControl, Select } from '../components';
-import useModalStore from '../store/useModalStore';
+import { useModalStore } from '../store';
 import { useLocation } from 'react-router-dom';
 import { MODAL_CONFIG } from '../components/modal/modalType';
 import { priceUnit, PriceUnitType } from '../constants/expenseTracker';
