+++ src/pages/MyPage.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 import { Link, useNavigate } from 'react-router-dom';
 import { Button, ShrinkMotionBlock } from '../components';
-import useUserStore from '../store/userStore';
+import { useUserStore } from '../store';
 import { navigationLinks, routes } from '../constants';
 
 const MyPage = () => {
