+++ src/pages/Register.tsx
@@ -5,11 +5,11 @@ import { Link, useNavigate } from 'react-router-dom';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { debounce } from 'es-toolkit';
 import supabase from '../supabase/service';
+import { useToastStore } from '../store';
 import { registerSchema, RegisterSchema } from '../components/auth/schema';
 import { Button, LabelInput, AuthLogo } from '../components';
 import { useFunnel, useLoading } from '../hooks';
 import { customPropReceiver, routes } from '../constants';
-import useToastStore from '../store/useToastStore';
 import { toastData } from '../constants/toast';
 import { addNewUser, isUserExist } from '../supabase/user';
 import AuthLayout from '../components/layout/AuthLayout';
