+++ src/pages/UpdateProfile.tsx
@@ -1,18 +1,17 @@
 import styled from '@emotion/styled';
 import { useNavigate } from 'react-router-dom';
 import { RiArrowRightSLine } from 'react-icons/ri';
+import { Session } from '@supabase/supabase-js';
+import { useQueryClient } from '@tanstack/react-query';
 import { Button, ShrinkMotionBlock } from '../components';
 import { format } from '../utils/date';
 import { useLoading } from '../hooks';
-import useToastStore from '../store/useToastStore';
-import useModalStore from '../store/useModalStore';
+import { useToastStore, useModalStore } from '../store';
 import { MODAL_CONFIG } from '../components/modal/modalType';
 import supabase from '../supabase/service';
-import useUserStore from '../store/userStore';
-import { useQueryClient } from '@tanstack/react-query';
+import { useUserStore } from '../store';
 import { toastData } from '../constants/toast';
 import { routes } from '../constants';
-import { Session } from '@supabase/supabase-js';
 
 const UpdateProfile = () => {
 	const queryClient = useQueryClient();
